+++ bitloops/src/api/tests/devql_repository_graph/projects.rs
@@ -1,0 +1,446 @@
+use super::*;
+
+#[tokio::test]
+async fn devql_project_queries_scope_paths_and_isolate_cross_project_resolution() {
+    let repo = seed_graphql_monorepo_repo();
+    let schema = crate::graphql::build_schema(crate::graphql::DevqlGraphqlContext::new(
+        repo.path().to_path_buf(),
+        super::super::super::db::DashboardDbPools::default(),
+    ));
+
+    let response = schema
+        .execute(async_graphql::Request::new(
+            r#"
+            {
+              repo(name: "demo") {
+                api: project(path: "packages/api") {
+                  path
+                  file(path: "src/caller.ts") {
+                    path
+                  }
+                  files(path: "src/*.ts") {
+                    path
+                  }
+                  artefacts(filter: { kind: FUNCTION }, first: 10) {
+                    totalCount
+                    edges {
+                      node {
+                        symbolFqn
+                        path
+                        outgoingDeps {
+                          totalCount
+                          edges {
+                            node {
+                              toSymbolRef
+                              toArtefact {
+                                symbolFqn
+                                path
+                              }
+                            }
+                          }
+                        }
+                      }
+                    }
+                  }
+                  deps(filter: { direction: OUT }, first: 10) {
+                    totalCount
+                    edges {
+                      node {
+                        toSymbolRef
+                        toArtefact {
+                          symbolFqn
+                          path
+                        }
+                      }
+                    }
+                  }
+                }
+                web: project(path: "packages/web") {
+                  path
+                  artefacts(filter: { kind: FUNCTION }, first: 10) {
+                    totalCount
+                    edges {
+                      node {
+                        symbolFqn
+                        path
+                      }
+                    }
+                  }
+                }
+              }
+            }
+            "#,
+        ))
+        .await;
+
+    assert!(
+        response.errors.is_empty(),
+        "graphql errors: {:?}",
+        response.errors
+    );
+
+    let json = response.data.into_json().expect("graphql data to json");
+    assert_eq!(json["repo"]["api"]["path"], "packages/api");
+    assert_eq!(
+        json["repo"]["api"]["file"]["path"],
+        "packages/api/src/caller.ts"
+    );
+    assert_eq!(
+        json["repo"]["api"]["files"],
+        json!([
+            { "path": "packages/api/src/caller.ts" },
+            { "path": "packages/api/src/target.ts" }
+        ])
+    );
+    assert_eq!(json["repo"]["api"]["artefacts"]["totalCount"], 2);
+    assert_eq!(
+        json["repo"]["api"]["artefacts"]["edges"][0]["node"]["symbolFqn"],
+        "packages/api/src/caller.ts::caller"
+    );
+    assert_eq!(
+        json["repo"]["api"]["artefacts"]["edges"][1]["node"]["symbolFqn"],
+        "packages/api/src/target.ts::target"
+    );
+    assert_eq!(
+        json["repo"]["api"]["artefacts"]["edges"][0]["node"]["outgoingDeps"]["totalCount"],
+        2
+    );
+    assert_eq!(
+        json["repo"]["api"]["artefacts"]["edges"][0]["node"]["outgoingDeps"]["edges"][0]["node"]["toArtefact"]
+            ["symbolFqn"],
+        "packages/api/src/target.ts::target"
+    );
+    assert_eq!(
+        json["repo"]["api"]["artefacts"]["edges"][0]["node"]["outgoingDeps"]["edges"][1]["node"]["toSymbolRef"],
+        "packages/web/src/page.ts::render"
+    );
+    assert_eq!(
+        json["repo"]["api"]["artefacts"]["edges"][0]["node"]["outgoingDeps"]["edges"][1]["node"]["toArtefact"],
+        serde_json::Value::Null
+    );
+    assert_eq!(json["repo"]["api"]["deps"]["totalCount"], 2);
+    assert_eq!(
+        json["repo"]["api"]["deps"]["edges"][1]["node"]["toArtefact"],
+        serde_json::Value::Null
+    );
+    assert_eq!(json["repo"]["web"]["path"], "packages/web");
+    assert_eq!(json["repo"]["web"]["artefacts"]["totalCount"], 1);
+    assert_eq!(
+        json["repo"]["web"]["artefacts"]["edges"][0]["node"]["symbolFqn"],
+        "packages/web/src/page.ts::render"
+    );
+}
+
+#[tokio::test]
+async fn devql_temporal_queries_resolve_historical_scope_once_and_propagate_to_children() {
+    let seeded = seed_graphql_temporal_repo();
+    let schema = crate::graphql::build_schema(crate::graphql::DevqlGraphqlContext::new(
+        seeded.repo.path().to_path_buf(),
+        super::super::super::db::DashboardDbPools::default(),
+    ));
+
+    let response = schema
+        .execute(async_graphql::Request::new(format!(
+            r#"
+            {{
+              repo(name: "demo") {{
+                repoScoped: asOf(input: {{ commit: "{}" }}) {{
+                  resolvedCommit
+                  project(path: "packages/api") {{
+                    path
+                    files(path: "src/*.ts") {{
+                      path
+                      blobSha
+                    }}
+                    file(path: "src/caller.ts") {{
+                      path
+                      artefacts(filter: {{ kind: FUNCTION }}, first: 10) {{
+                        totalCount
+                        edges {{
+                          node {{
+                            symbolFqn
+                            outgoingDeps {{
+                              totalCount
+                              edges {{
+                                node {{
+                                  toArtefact {{
+                                    symbolFqn
+                                    path
+                                  }}
+                                }}
+                              }}
+                            }}
+                          }}
+                        }}
+                      }}
+                    }}
+                    artefacts(filter: {{ kind: FUNCTION }}, first: 10) {{
+                      totalCount
+                      edges {{
+                        node {{
+                          symbolFqn
+                          path
+                        }}
+                      }}
+                    }}
+                    deps(filter: {{ direction: OUT }}, first: 10) {{
+                      totalCount
+                      edges {{
+                        node {{
+                          toArtefact {{
+                            symbolFqn
+                            path
+                          }}
+                        }}
+                      }}
+                    }}
+                  }}
+                }}
+                project(path: "packages/api") {{
+                  projectScoped: asOf(input: {{ commit: "{}" }}) {{
+                    resolvedCommit
+                    artefacts(filter: {{ kind: FUNCTION }}, first: 10) {{
+                      totalCount
+                      edges {{
+                        node {{
+                          symbolFqn
+                          path
+                        }}
+                      }}
+                    }}
+                  }}
+                }}
+              }}
+            }}
+            "#,
+            seeded.first_commit, seeded.first_commit,
+        )))
+        .await;
+
+    assert!(
+        response.errors.is_empty(),
+        "graphql errors: {:?}",
+        response.errors
+    );
+
+    let json = response.data.into_json().expect("graphql data to json");
+    assert_eq!(
+        json["repo"]["repoScoped"]["resolvedCommit"],
+        seeded.first_commit
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["path"],
+        "packages/api"
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["files"],
+        json!([
+            {
+                "path": "packages/api/src/caller.ts",
+                "blobSha": "blob-api-caller-v1"
+            },
+            {
+                "path": "packages/api/src/target.ts",
+                "blobSha": "blob-api-target-v1"
+            }
+        ])
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["file"]["path"],
+        "packages/api/src/caller.ts"
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["artefacts"]["totalCount"],
+        2
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["artefacts"]["edges"][0]["node"]["symbolFqn"],
+        "packages/api/src/caller.ts::caller"
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["artefacts"]["edges"][1]["node"]["symbolFqn"],
+        "packages/api/src/target.ts::target"
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["file"]["artefacts"]["totalCount"],
+        1
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["file"]["artefacts"]["edges"][0]["node"]["symbolFqn"],
+        "packages/api/src/caller.ts::caller"
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["file"]["artefacts"]["edges"][0]["node"]["outgoingDeps"]
+            ["totalCount"],
+        1
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["file"]["artefacts"]["edges"][0]["node"]["outgoingDeps"]
+            ["edges"][0]["node"]["toArtefact"]["symbolFqn"],
+        "packages/api/src/target.ts::target"
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["deps"]["totalCount"],
+        1
+    );
+    assert_eq!(
+        json["repo"]["repoScoped"]["project"]["deps"]["edges"][0]["node"]["toArtefact"]["symbolFqn"],
+        "packages/api/src/target.ts::target"
+    );
+    assert_eq!(
+        json["repo"]["project"]["projectScoped"]["resolvedCommit"],
+        seeded.first_commit
+    );
+    assert_eq!(
+        json["repo"]["project"]["projectScoped"]["artefacts"]["totalCount"],
+        2
+    );
+}
+
+#[tokio::test]
+async fn devql_temporal_queries_validate_inputs_and_unknown_refs() {
+    let seeded = seed_graphql_temporal_repo();
+    let schema = crate::graphql::build_schema(crate::graphql::DevqlGraphqlContext::new(
+        seeded.repo.path().to_path_buf(),
+        super::super::super::db::DashboardDbPools::default(),
+    ));
+
+    let invalid_selector = schema
+        .execute(async_graphql::Request::new(
+            r#"
+            {
+              repo(name: "demo") {
+                asOf(input: { commit: "abc123", ref: "main" }) {
+                  resolvedCommit
+                }
+              }
+            }
+            "#,
+        ))
+        .await;
+    assert_eq!(
+        invalid_selector.errors.len(),
+        1,
+        "expected invalid asOf selector error"
+    );
+    assert_eq!(
+        invalid_selector.errors[0]
+            .extensions
+            .as_ref()
+            .and_then(|extensions| extensions.get("code")),
+        Some(&async_graphql::Value::from("BAD_USER_INPUT"))
+    );
+
+    let unknown_ref = schema
+        .execute(async_graphql::Request::new(
+            r#"
+            {
+              repo(name: "demo") {
+                asOf(input: { ref: "refs/heads/missing-temporal-branch" }) {
+                  resolvedCommit
+                }
+              }
+            }
+            "#,
+        ))
+        .await;
+    assert_eq!(
+        unknown_ref.errors.len(),
+        1,
+        "expected one unknown-ref error"
+    );
+    assert_eq!(
+        unknown_ref.errors[0]
+            .extensions
+            .as_ref()
+            .and_then(|extensions| extensions.get("code")),
+        Some(&async_graphql::Value::from("BAD_USER_INPUT"))
+    );
+}
+
+#[tokio::test]
+async fn devql_project_queries_validate_project_paths() {
+    let repo = seed_graphql_monorepo_repo();
+    let schema = crate::graphql::build_schema(crate::graphql::DevqlGraphqlContext::new(
+        repo.path().to_path_buf(),
+        super::super::super::db::DashboardDbPools::default(),
+    ));
+
+    let invalid = schema
+        .execute(async_graphql::Request::new(
+            r#"
+            {
+              repo(name: "demo") {
+                project(path: "../packages/api") {
+                  path
+                }
+              }
+            }
+            "#,
+        ))
+        .await;
+    assert_eq!(
+        invalid.errors.len(),
+        1,
+        "expected invalid project path error"
+    );
+    assert_eq!(
+        invalid.errors[0]
+            .extensions
+            .as_ref()
+            .and_then(|extensions| extensions.get("code")),
+        Some(&async_graphql::Value::from("BAD_USER_INPUT"))
+    );
+
+    let missing = schema
+        .execute(async_graphql::Request::new(
+            r#"
+            {
+              repo(name: "demo") {
+                project(path: "packages/missing") {
+                  path
+                }
+              }
+            }
+            "#,
+        ))
+        .await;
+    assert_eq!(
+        missing.errors.len(),
+        1,
+        "expected missing project path error"
+    );
+    assert_eq!(
+        missing.errors[0]
+            .extensions
+            .as_ref()
+            .and_then(|extensions| extensions.get("code")),
+        Some(&async_graphql::Value::from("BAD_USER_INPUT"))
+    );
+
+    let not_directory = schema
+        .execute(async_graphql::Request::new(
+            r#"
+            {
+              repo(name: "demo") {
+                project(path: "packages/api/src/caller.ts") {
+                  path
+                }
+              }
+            }
+            "#,
+        ))
+        .await;
+    assert_eq!(
+        not_directory.errors.len(),
+        1,
+        "expected non-directory project path error"
+    );
+    assert_eq!(
+        not_directory.errors[0]
+            .extensions
+            .as_ref()
+            .and_then(|extensions| extensions.get("code")),
+        Some(&async_graphql::Value::from("BAD_USER_INPUT"))
+    );
+}
